# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ایده‌ها",
        "description": "API برای مدیریت ایده‌ها",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "سرور اصلی"
        }
    ],
    "paths": {
...
        "/api/v1/admin/topic-tags": {
            "get": {
                "tags": [
                    "Admin - topic-tags"
                ],
                "summary": "لیست ارتباط موضوع و تگ",
                "description": "نمایش لیستی از ارتباط‌های بین موضوعات و تگ‌ها با قابلیت فیلتر بر اساس topic_id یا tag_id",
                "operationId": "topicTagsList",
                "parameters": [
                    {
                        "name": "topic_id",
                        "in": "query",
                        "description": "فیلتر بر اساس آیدی موضوع",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "tag_id",
                        "in": "query",
                        "description": "فیلتر بر اساس آیدی تگ",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "topic_tags": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/TopicTagsIndexResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "خطای سرور",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "خطای داخلی سرور"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin - topic-tags"
                ],
                "summary": "ایجاد ارتباط موضوع و تگ",
                "description": "ایجاد رکورد جدید در جدول واسط topic_tags جهت اتصال یک موضوع به یک تگ خاص",
                "operationId": "createTopicTag",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "topic_id",
                                    "tag_id"
                                ],
                                "properties": {
                                    "topic_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tag_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 101
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "خطای اعتبارسنجی",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "خطا در اعتبارسنجی داده‌ها"
                                        },
                                        "errors": {
                                            "properties": {
                                                "topic_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "شناسه موضوع الزامی است"
                                                    }
                                                },
                                                "tag_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "شناسه تگ الزامی است"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "خطای سرور",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "خطای داخلی سرور"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TopicTagsIndexResource": {
                "properties": {
                    "topic_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tag_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date",
                        "example": "1402/01/01"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your JWT Bearer token in the format **Bearer {your token}**"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}